--- a/doc/vector.xxml
+++ b/doc/vector.xxml
@@ -136,6 +136,7 @@
 
 <section id="set-operations-on-sorted-vectors"><title>Set operations on sorted vectors</title>
 <!-- doxrox-include igraph_vector_intersect_sorted -->
+<!-- doxrox-include igraph_vector_intersection_size_sorted -->
 <!-- doxrox-include igraph_vector_difference_sorted -->
 </section>