--- a/tests/CMakeLists.txt
+++ b/tests/CMakeLists.txt
@@ -1021,6 +1021,7 @@ add_benchmarks(
   igraph_voronoi
   inc_vs_adj
   lad
+  intersection
 )
 
 # dot product