… tests/unit/igraph_ecc.out
… tests/unit/vector2.out