Benchmarking for #2619, ~~not to be merged~~. **EDIT: Refactored into an improvement, ready for merging.**

----

This PR implements an intersection size calculation using the Baeza-Yates algorithm (we already have a function to compute the actual intersection using this algorithm), then tries to use it in `igraph_ecc()`, switching between this and a linear scan based  on the ratio of input vector sizes.

These benchmarks are indirect, through `igraph_ecc()`, which heavily relies on intersection size computations, specifically for counting common neighbours. The existing `igraph_ecc()` benchmark testa both Erdos-Renyi graphs, where neighbour lists have similar sizes, and  Barabasi graphs, where neighbour list sizes can span several orders of magnitude. It is the latter case where Baeza-Yates makes sense.

### Baseline performance, linear scan only

```
|> Benchmark file: /Users/szhorvat/Repos/igraph-main/igraph/tests/benchmarks/igraph_ecc.c
Erdos-Renyi GNM:

|  1 vc=100, ec=5000, k=3, all, 100x                                               0.146s  0.144s      0s
|  2 vc=100, ec=5000, k=3, subset: all, 100x                                       0.115s  0.112s      0s
|  3 vc=100, ec=5000, k=3, subset: 100, 1000x                                      0.058s  0.057s      0s
|  4 vc=100, ec=5000, k=4, all, 100x                                                7.72s   7.63s  0.002s
|  4 vc=100, ec=5000, k=4, subset: all, 100x                                        7.72s   7.65s  0.002s
|  5 vc=100, ec=5000, k=4, subset: 100, 1000x                                        1.6s   1.58s      0s

|  1 vc=1000, ec=5000, k=3, all, 100x                                              0.037s  0.036s      0s
|  2 vc=1000, ec=5000, k=3, subset: all, 100x                                      0.044s  0.042s      0s
|  3 vc=1000, ec=5000, k=3, subset: 100, 1000x                                     0.024s  0.023s      0s
|  4 vc=1000, ec=5000, k=4, all, 100x                                              0.249s  0.245s      0s
|  4 vc=1000, ec=5000, k=4, subset: all, 100x                                       0.26s  0.258s  0.001s
|  5 vc=1000, ec=5000, k=4, subset: 100, 1000x                                     0.119s  0.117s  0.001s

|  1 vc=10000, ec=10000, k=3, all, 100x                                            0.065s  0.064s  0.001s
|  2 vc=10000, ec=10000, k=3, subset: all, 100x                                    0.093s   0.09s  0.002s
|  3 vc=10000, ec=10000, k=3, subset: 100, 1000x                                   0.021s  0.021s      0s
|  4 vc=10000, ec=10000, k=4, all, 100x                                            0.085s  0.083s  0.001s
|  4 vc=10000, ec=10000, k=4, subset: all, 100x                                    0.113s   0.11s  0.002s
|  5 vc=10000, ec=10000, k=4, subset: 100, 1000x                                   0.028s  0.028s      0s

|  1 vc=10000, ec=50000, k=3, all, 100x                                             0.43s  0.424s  0.004s
|  2 vc=10000, ec=50000, k=3, subset: all, 100x                                    0.488s   0.48s  0.006s
|  3 vc=10000, ec=50000, k=3, subset: 100, 1000x                                   0.031s  0.031s      0s
|  4 vc=10000, ec=50000, k=4, all, 100x                                             2.79s   2.77s  0.005s
|  4 vc=10000, ec=50000, k=4, subset: all, 100x                                     2.94s   2.91s  0.006s
|  5 vc=10000, ec=50000, k=4, subset: 100, 1000x                                   0.199s  0.194s  0.003s

Barabasi:

|  1 vc=10000, ec=50000, k=3, all, 100x                                            0.862s  0.856s  0.003s
|  2 vc=10000, ec=50000, k=3, subset: all, 100x                                    0.932s  0.923s  0.005s
|  3 vc=10000, ec=50000, k=3, subset: 100, 1000x                                   0.039s  0.039s      0s
|  4 vc=10000, ec=50000, k=4, all, 100x                                             13.5s   13.5s  0.005s
|  4 vc=10000, ec=50000, k=4, subset: all, 100x                                     13.6s   13.6s  0.008s
|  5 vc=10000, ec=50000, k=4, subset: 100, 1000x                                   0.288s  0.284s  0.001s
```

### Using Baeza-Yates unconditionally

Notice that timings for the Erdos-Renyi case suffer significantly compared to the baseline, but the performance on scale-free (Barabasi) networks stays similar.

```
|> Benchmark file: /Users/szhorvat/Repos/igraph-main/igraph/tests/benchmarks/igraph_ecc.c
Erdos-Renyi GNM:

|  1 vc=100, ec=5000, k=3, all, 100x                                               0.398s  0.392s      0s
|  2 vc=100, ec=5000, k=3, subset: all, 100x                                       0.373s  0.363s      0s
|  3 vc=100, ec=5000, k=3, subset: 100, 1000x                                      0.114s  0.112s      0s
|  4 vc=100, ec=5000, k=4, all, 100x                                                21.7s   21.6s  0.003s
|  4 vc=100, ec=5000, k=4, subset: all, 100x                                        21.8s   21.7s  0.003s
|  5 vc=100, ec=5000, k=4, subset: 100, 1000x                                        5.1s   5.05s  0.001s

|  1 vc=1000, ec=5000, k=3, all, 100x                                              0.068s  0.068s      0s
|  2 vc=1000, ec=5000, k=3, subset: all, 100x                                      0.072s  0.071s      0s
|  3 vc=1000, ec=5000, k=3, subset: 100, 1000x                                     0.026s  0.025s      0s
|  4 vc=1000, ec=5000, k=4, all, 100x                                              0.572s  0.564s      0s
|  4 vc=1000, ec=5000, k=4, subset: all, 100x                                      0.577s  0.571s  0.001s
|  5 vc=1000, ec=5000, k=4, subset: 100, 1000x                                     0.168s  0.166s  0.001s

|  1 vc=10000, ec=10000, k=3, all, 100x                                            0.078s  0.076s  0.001s
|  2 vc=10000, ec=10000, k=3, subset: all, 100x                                    0.105s  0.102s  0.002s
|  3 vc=10000, ec=10000, k=3, subset: 100, 1000x                                   0.021s  0.021s      0s
|  4 vc=10000, ec=10000, k=4, all, 100x                                            0.109s  0.107s  0.001s
|  4 vc=10000, ec=10000, k=4, subset: all, 100x                                    0.135s  0.132s  0.002s
|  5 vc=10000, ec=10000, k=4, subset: 100, 1000x                                   0.029s  0.029s      0s

|  1 vc=10000, ec=50000, k=3, all, 100x                                            0.729s  0.721s  0.004s
|  2 vc=10000, ec=50000, k=3, subset: all, 100x                                    0.787s  0.778s  0.006s
|  3 vc=10000, ec=50000, k=3, subset: 100, 1000x                                   0.032s  0.032s      0s
|  4 vc=10000, ec=50000, k=4, all, 100x                                             5.84s   5.82s  0.005s
|  4 vc=10000, ec=50000, k=4, subset: all, 100x                                     5.97s   5.94s  0.007s
|  5 vc=10000, ec=50000, k=4, subset: 100, 1000x                                   0.253s  0.246s  0.003s

Barabasi:

|  1 vc=10000, ec=50000, k=3, all, 100x                                            0.975s  0.963s  0.003s
|  2 vc=10000, ec=50000, k=3, subset: all, 100x                                     1.04s   1.02s  0.005s
|  3 vc=10000, ec=50000, k=3, subset: 100, 1000x                                   0.033s  0.033s      0s
|  4 vc=10000, ec=50000, k=4, all, 100x                                             13.3s   13.3s  0.006s
|  4 vc=10000, ec=50000, k=4, subset: all, 100x                                     13.4s   13.3s  0.008s
|  5 vc=10000, ec=50000, k=4, subset: 100, 1000x                                   0.311s  0.305s  0.001s
```

### Switch to Baeza-Yates when the input size ratio exceeds 15

The performance on Erdos-Renyi is similar to the baseline. The performance on Barabasi is improved.  The results are about the same for a ratio threshold between 10-15. Above 20 or below the performance gains start to disappear.

This benchmark was done on an Apple M2.

In the original Baeza-Yates paper, Fig 3 seems to suggest an optimum switching threshold of about 30.

```
|> Benchmark file: /Users/szhorvat/Repos/igraph-main/igraph/tests/benchmarks/igraph_ecc.c
Erdos-Renyi GNM:

|  1 vc=100, ec=5000, k=3, all, 100x                                               0.147s  0.144s      0s
|  2 vc=100, ec=5000, k=3, subset: all, 100x                                       0.115s  0.113s      0s
|  3 vc=100, ec=5000, k=3, subset: 100, 1000x                                      0.069s  0.069s      0s
|  4 vc=100, ec=5000, k=4, all, 100x                                                 7.7s   7.65s  0.002s
|  4 vc=100, ec=5000, k=4, subset: all, 100x                                        7.69s   7.65s  0.002s
|  5 vc=100, ec=5000, k=4, subset: 100, 1000x                                       1.62s    1.6s      0s

|  1 vc=1000, ec=5000, k=3, all, 100x                                              0.037s  0.036s      0s
|  2 vc=1000, ec=5000, k=3, subset: all, 100x                                      0.044s  0.042s      0s
|  3 vc=1000, ec=5000, k=3, subset: 100, 1000x                                     0.023s  0.023s      0s
|  4 vc=1000, ec=5000, k=4, all, 100x                                              0.253s   0.25s      0s
|  4 vc=1000, ec=5000, k=4, subset: all, 100x                                      0.262s   0.26s      0s
|  5 vc=1000, ec=5000, k=4, subset: 100, 1000x                                      0.12s  0.118s  0.001s

|  1 vc=10000, ec=10000, k=3, all, 100x                                            0.066s  0.065s  0.001s
|  2 vc=10000, ec=10000, k=3, subset: all, 100x                                    0.094s  0.091s  0.002s
|  3 vc=10000, ec=10000, k=3, subset: 100, 1000x                                   0.021s  0.021s      0s
|  4 vc=10000, ec=10000, k=4, all, 100x                                            0.086s  0.085s  0.001s
|  4 vc=10000, ec=10000, k=4, subset: all, 100x                                    0.113s  0.111s  0.002s
|  5 vc=10000, ec=10000, k=4, subset: 100, 1000x                                   0.028s  0.028s      0s

|  1 vc=10000, ec=50000, k=3, all, 100x                                            0.428s  0.423s  0.004s
|  2 vc=10000, ec=50000, k=3, subset: all, 100x                                    0.487s  0.479s  0.006s
|  3 vc=10000, ec=50000, k=3, subset: 100, 1000x                                   0.032s  0.032s      0s
|  4 vc=10000, ec=50000, k=4, all, 100x                                             2.83s   2.81s  0.005s
|  4 vc=10000, ec=50000, k=4, subset: all, 100x                                     2.95s   2.92s  0.006s
|  5 vc=10000, ec=50000, k=4, subset: 100, 1000x                                   0.198s  0.192s  0.004s

Barabasi:

|  1 vc=10000, ec=50000, k=3, all, 100x                                            0.652s  0.645s  0.003s
|  2 vc=10000, ec=50000, k=3, subset: all, 100x                                    0.713s  0.706s  0.005s
|  3 vc=10000, ec=50000, k=3, subset: 100, 1000x                                   0.036s  0.036s      0s
|  4 vc=10000, ec=50000, k=4, all, 100x                                              8.4s   8.36s  0.005s
|  4 vc=10000, ec=50000, k=4, subset: all, 100x                                     8.47s   8.42s  0.007s
|  5 vc=10000, ec=50000, k=4, subset: 100, 1000x                                   0.221s  0.216s  0.001s
```