@szhorvat 2024-05-26
This benchmarking is not sufficient for making a decision on when to switch algorithms, but it does illustrate what practical gains we might expect. A direct benchmarking of the intersection function will be necessary.

What is seem clear though is that for practical use cases, an unconditional Baeza-Yates is a worse choice than a naive linear scan. So the set intersection function _should_ be updated, probably in a conservative manner (switch to Baeza-Yates less eagerly).