@szhorvat 2024-05-26
This shows the timing ratio of the linear scan/merge to Baeza-Yates as a function of the smaller vector length, $n$, and the ratio of the larger to smaller vector lengths, $r$. Note the log scale.

<img width="360" alt="image" src="https://github.com/igraph/igraph/assets/1212871/7c5b9ce2-1214-4ec2-bf7d-3fdb8e4cab9c">

The dashed line corresponds to a ratio of 1, i.e. the two methods performing equally. Contours are spaced at ratios of $\sqrt{e}$ apart.

What we see is that for small lists Baeza-Yates is worth it above a length ratio of 5, and for large ones ($n > 1000$) a ratio of 30. No doubt this is some sort of cache effect that does not generalize to other machines. Thus I'm still now sure how to pick the best method for given input sizes.