@szhorvat 2024-05-26
I pushed a version which seems to work well in practice on my machine, and uses a threshold of $r=10$. I implemented a branchless (i.e. free of conditionals) version of the linear scan, which gave a further speedup. This is what the ECC benchmark looks like now on my machine:

```
|> Benchmark file: /Users/szhorvat/Repos/igraph-main/igraph/tests/benchmarks/igraph_ecc.c
Erdos-Renyi GNM:

|  1 vc=100, ec=5000, k=3, all, 100x                                               0.133s  0.129s      0s
|  2 vc=100, ec=5000, k=3, subset: all, 100x                                       0.102s  0.099s      0s
|  3 vc=100, ec=5000, k=3, subset: 100, 1000x                                      0.069s  0.067s      0s
|  4 vc=100, ec=5000, k=4, all, 100x                                                6.63s   6.53s  0.001s
|  4 vc=100, ec=5000, k=4, subset: all, 100x                                        6.61s    6.5s  0.001s
|  5 vc=100, ec=5000, k=4, subset: 100, 1000x                                        1.4s   1.36s      0s

|  1 vc=1000, ec=5000, k=3, all, 100x                                              0.033s  0.032s      0s
|  2 vc=1000, ec=5000, k=3, subset: all, 100x                                       0.04s  0.039s      0s
|  3 vc=1000, ec=5000, k=3, subset: 100, 1000x                                     0.023s  0.023s      0s
|  4 vc=1000, ec=5000, k=4, all, 100x                                               0.22s  0.215s      0s
|  4 vc=1000, ec=5000, k=4, subset: all, 100x                                       0.23s  0.225s  0.001s
|  5 vc=1000, ec=5000, k=4, subset: 100, 1000x                                     0.114s  0.111s  0.001s

|  1 vc=10000, ec=10000, k=3, all, 100x                                            0.062s   0.06s  0.001s
|  2 vc=10000, ec=10000, k=3, subset: all, 100x                                    0.092s  0.089s  0.002s
|  3 vc=10000, ec=10000, k=3, subset: 100, 1000x                                   0.022s  0.021s      0s
|  4 vc=10000, ec=10000, k=4, all, 100x                                            0.082s  0.079s  0.001s
|  4 vc=10000, ec=10000, k=4, subset: all, 100x                                    0.112s  0.107s  0.002s
|  5 vc=10000, ec=10000, k=4, subset: 100, 1000x                                   0.028s  0.027s      0s

|  1 vc=10000, ec=50000, k=3, all, 100x                                            0.402s   0.39s  0.004s
|  2 vc=10000, ec=50000, k=3, subset: all, 100x                                    0.457s  0.443s  0.006s
|  3 vc=10000, ec=50000, k=3, subset: 100, 1000x                                   0.031s  0.031s      0s
|  4 vc=10000, ec=50000, k=4, all, 100x                                             2.47s   2.43s  0.005s
|  4 vc=10000, ec=50000, k=4, subset: all, 100x                                     2.59s   2.55s  0.006s
|  5 vc=10000, ec=50000, k=4, subset: 100, 1000x                                   0.193s  0.187s  0.003s

Barabasi:

|  1 vc=10000, ec=50000, k=3, all, 100x                                            0.614s  0.601s  0.003s
|  2 vc=10000, ec=50000, k=3, subset: all, 100x                                    0.679s  0.663s  0.005s
|  3 vc=10000, ec=50000, k=3, subset: 100, 1000x                                   0.035s  0.034s      0s
|  4 vc=10000, ec=50000, k=4, all, 100x                                             7.79s   7.69s  0.003s
|  4 vc=10000, ec=50000, k=4, subset: all, 100x                                     7.87s   7.77s  0.006s
|  5 vc=10000, ec=50000, k=4, subset: 100, 1000x                                   0.201s  0.197s  0.001s
```