@szhorvat 2024-05-26
@ntamas What do you think? Shall we go ahead with a simple vector length ratio test?  Note that this PR is still not cleaned up for merging!