@ntamas 2024-05-27
> Shall we go ahead with a simple vector length ratio test?

Sounds good to me.