@szhorvat 2024-05-27
Benchmark on RPi 5, 64-bit:

<img width="360" alt="image" src="https://github.com/igraph/igraph/assets/1212871/1489ec59-4444-4abe-9d8d-59c4ed030897">

Benchmark on RPi 4, 32-bit:

<img width="360" alt="image" src="https://github.com/igraph/igraph/assets/1212871/cb18371c-45bc-4741-b217-13f5b98f0d31">